New logging mod registration
Major changes in this PR are:
1. Removed the SISL_LOGGING_* macro and use the macro level check with string
literal templates instead of dsym check. dsym doesn't work when compiled with
multiple libraries. Moreover the requirement to declare and use a module is
greatly simplified

2. Make the cache destroy to call the evictor during destruction for each
hash node. At present if we have multiple caches on same evictor and if one
cache is destroyed, it will leave with evictor having some dead nodes inside.
