+++ include/sisl/grpc/rpc_call.hpp
@@ -29,8 +29,6 @@
 #include <sisl/utility/enum.hpp>
 #include "rpc_common.hpp"
 
-SISL_LOGGING_DECL(grpc_server)
-
 #define RPC_SERVER_LOG(level, msg, ...)                                                                                \
     LOG##level##MOD_FMT(grpc_server, ([&](fmt::memory_buffer& buf, const char* __m, auto&&... args) -> bool {          \
                             fmt::vformat_to(fmt::appender{buf}, std::string_view{"[{}:{}] [RPC={} id={}] "},           \
