+++ src/cache/tests/test_range_hashmap.cpp
@@ -26,8 +26,6 @@
 #include <sisl/cache/range_hashmap.hpp>
 
 using namespace sisl;
-REGISTER_LOG_MOD(test_hashmap)
-
 static uint32_t g_max_offset;
 static constexpr uint32_t per_val_size = 128;
