+++ src/cache/tests/test_simple_cache.cpp
@@ -33,8 +33,6 @@
 #include <sisl/cache/lru_evictor.hpp>
 
 using namespace sisl;
-REGISTER_LOG_MOD(test_simplecache)
-
 static constexpr uint32_t g_val_size{512};
 static thread_local std::random_device g_rd{};
 static thread_local std::default_random_engine g_re{g_rd()};
