+++ src/fds/tests/test_jemalloc_helper.cpp
@@ -33,8 +33,6 @@
 
 using namespace sisl;
 
-SISL_LOGGING_INIT(test_jemalloc)
-
 namespace {
 uint32_t g_num_threads;
