+++ src/fds/tests/test_stream_tracker.cpp
@@ -26,8 +26,6 @@
 
 using namespace sisl;
 
-SISL_LOGGING_INIT(test_stream_tracker)
-
 namespace {
 struct TestData {
     TestData(int val) : m_value(val) {}
