+++ src/metrics/tests/functionality_test.cpp
@@ -33,7 +33,6 @@ constexpr size_t ITERATIONS{2};
 // CREATE_REPORT;
 THREAD_BUFFER_INIT
 RCU_REGISTER_INIT
-SISL_LOGGING_INIT(vmod_metrics_framework)
 
 using namespace sisl;
