+++ src/metrics/tests/wrapper_test.cpp
@@ -24,8 +24,6 @@
 
 #include "sisl/metrics/metrics.hpp"
 
-SISL_LOGGING_INIT(vmod_metrics_framework)
-
 RCU_REGISTER_INIT
 
 using namespace sisl;
