+++ src/version/tests/test_version.cpp
@@ -10,7 +10,6 @@
 using namespace sisl;
 
 SISL_OPTIONS_ENABLE(logging)
-SISL_LOGGING_INIT()
 
 void entry() {
     auto ver{version::Semver200_version(BOOST_PP_STRINGIZE(PACKAGE_VERSION))};
